removed the redundant signup screen and made the login link on register logout you out and take you back to the login page
