+++ static/js/authRedirect.js
@@ -76,19 +76,24 @@ function signIn() {
 
 function signOut() {
 
-    /**
-     * You can pass a custom request object below. This will override the initial configuration. For more information, visit:
-     * https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-browser/docs/request-response-object.md#request
-     */
+    // console.log("hello world?")
+    // /**
+    //  * You can pass a custom request object below. This will override the initial configuration. For more information, visit:
+    //  * https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-browser/docs/request-response-object.md#request
+    //  */
 
-    // Choose which account to logout from by passing a username.
-    const logoutRequest = {
-        account: myMSALObj.getAccountByUsername(username),
-        // postLogoutRedirectUri: 'http://localhost:3000/signout', // Simply remove this line if you would like navigate to index page after logout.
+    // // Choose which account to logout from by passing a username.
+    // const logoutRequest = {
+    //     account: myMSALObj.getAccountByUsername(username),
+    //     // postLogoutRedirectUri: 'http://localhost:3000/signout', // Simply remove this line if you would like navigate to index page after logout.
 
-    };
+    // };
 
-    myMSALObj.logoutRedirect(logoutRequest);
+    // myMSALObj.logoutRedirect(logoutRequest);
+
+
+
+    logout()
 }
 
 function register() {
