+++ templates/signin-oidc.html
@@ -54,7 +54,7 @@
                     </form>
 
                     <div class="form-link">
-                        <span>Don't have an account? <a href="#" class="link signup-link">Signup</a></span>
+                        <span>Don't have an account? <a onclick="signIn()" class="signup-link">Signup</a></span>
                     </div>
                 </div>
 
@@ -145,7 +145,7 @@
                     </form>
 
                     <div class="form-link">
-                        <span>Already have an account? <a href="#" class="link login-link">Login</a></span>
+                        <span>Already have an account? <a onclick="signOut()" class="link login-link">Login</a></span>
                     </div>
                 </div>
             </div>
